fix: extract ShowCardOverlay component from GlobalMap
renderShowCard was a render function living inside GlobalMap, the
same pattern we moved away from with GlobalMapHeader. Pull it into
its own component so GlobalMap only wires state to it.
